requirements.txt
transformers==4.38.1
accelerate==0.27.2
einops
icecream
ipywidgets
flask
uvicorn
fastapi
markdown2
gradio
sconf
tensorboardX==2.5
tensorboard
h5py
sentencepiece
peft
opencv-python
decord
chardet
cchardet
prettytable
tqdm
flash-attn==2.3.2
protobuf==3.19.6
scikit-learn
omegaconf
hydra-core
deepspeed==0.10.0
braceexpand
webdataset==0.2.48
Levenshtein
openai==0.28
jsonlines
timm==0.9.8
open_clip_torch==2.23.0
pytorch-lightning==2.0.6
